restrictions: create single_table_query_restrictions in analyzed_where_clause
single_table_query_restrictions is the new representation
used to analyze single table queries.

It's created by giving it the WHERE clause and table schema.
It splits the WHERE clause into three parts:
* restrictions containing only partition key columns (e.g. p = 1, p1 < p2)
* restrictions containing only clustering key columns (e.g. c1 < 2, udf(c1, c2))
* All other restrictions (e.g. TRUE, my_func(pk, ck), ...)

Then it creates partition_key_restrictions and
clustering_key_restrictions, which are given these parts.

The idea is that partition_key_restrictions contains all data
which can be used to choose a partition. In the future
we could allow to do things like 'p1 = ? and p2 = ? and p1 < p2"
without filtering. Code able to handle this would be put
in partition_key_restrictions.

The WHERE clause has to be prepared before passing it
to the constructor. statement_restrictions has much more
complicated code to prepare the WHERE because it does some
additional validation checks on the binary operators.
This validation has already been done by statement_restrictions
constructor so we can just use prepare_expression.

Once grammar is fully relaxed we should be able to get rid
of all these validation checks. A correct expression that can be
prepared should also mean a correct query can be executed,
although maybe with filtering.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>