gms: feature_service: pass `db:config&` directly to `feature_service`
Previously `feature_service` constructor obtained a `feature_config`
struct which used `db::config&` to calculate the set of disabled
features. `feature_config` could also disable some extra features which
was used for testing.

In later commits we want `feature_service` to observe `db::config` as it
changes, so the Scylla node can start supporting new features in runtime
(through a config reload). The `feature_config` abstraction makes this
awkward to do, as it was designed to be calculated one time from
`db::config` and passed to `feature_service` without a `db::config`
reference.

Get rid of `feature_config` and pass a `db::config&` reference directly
to `feature_service`, which calculates the set of disabled features
internally. In the following commits it will also observe the config for
runtime changes. To handle additional disabled features for testing
purposes, we replace `feature_config` with a `custom_feature_config_for_tests`
struct which only contains the features disabled by the test that uses
it.