mutation_compactor: add validator
The mutation compactor is used on most read-paths we have, so adding a
validator to it gives us a good coverage, in particular it gives us full
coverage of queries and compaction.
The validator validates mutation fragment kind monotonicity as that is
quite cheap. We should probably make this configurable at some point.
The addition of this validator should prevent the worst of the fragment
reordering bugs to affect reads/compaction.