test.py: don't refresh node list on topology change
Due to timeouts seen when a host was down, we started doing a node
refresh after each topology change.

But this is no longer necessary after the recent timeout fixes and node
refresh is causing an issue in CI when the driver is doing other things
and it's closed.

So we can safely remove these calls to control_connection's
refresh_node_list_and_token_map().

Signed-off-by: Alejo Sanchez <alejo.sanchez@scylladb.com>