raft: (address_map) speed up initial cluster assembly
Even though we boot the initial address map with
the persistent state and subscribe to gossip updates,
this is not enough to make the approach with getting
IP address updates from gossip even barely work.

Without changes introduced in this patch, which should be
unnecessary, the tests don't pass.

Case 1: we have a cluster with server {A} which is a
leader, and {B} joins {A}. {A} can't push a snapshot
to {B} because {A} doesn't know {B}'s ip address.

This patch fixes it by taking B's IP address from
group0_modify_config RPC.

Case 2: we have a cluster with server {A} and {B}, A is a
leader, and {C} send modify_config request to {B},
which then forwards it to {A}. Once again, {A} doesn't
know {B}'s IP address.