cql3: expr: implement is_satisfied_by using evaluate
is_satisfied_by takes an expression and
checks if that expression is true or not.

This can now be implemented using evaluate().
Until recently evaluate() wasn't able to handle
certain types of expressions like binary_operator
or conjunction, but now it has gained this ability
and is_satisfied_by can be reduced to a few lines
of code.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>