test/pylib: scylla_cluster: don't uninstall servers removed by `server_remove`
`ScyllaCluster` has a `server_remove` function which shuts down the
server and 'removes' it from the `ScyllaCluster`'s set of servers so it
cannot be brought back again by the cluster and the test that uses it.

However, it does not actually remove the server from the Scylla cluster.
It only removes it from the test framework and shuts down the
corresponding Scylla node. The other nodes in the cluster know about the
server.

`server_remove` would `uninstall()` the `ScyllaServer`, which caused its
IP to be released into `host_registry`, which means it could be reused
by other clusters in other tests.

It could happen that after a test 'removed' a server (with
`server_remove`), another test would start with a cluster that reused
this server's IP; the tests would then run concurrently. The cluster
from the old test would then contact the new server with the old IP in
the new cluster, and weird things would happen. The driver session from
the old test would contact the new server, considering it to be a part
of the cluster. This would cause, for example, the test to fail on
cleanup when it tried to drop the test keyspace - if it used the new
server with the old IP to do that.

Another effect of `uninstall()` in `server_remove` was losing the
removed server's logs and work directories. If the test that did
`server_remove` failed, these logs and workdirs could be useful for
analysis.

We fix all this by not `uninstall`ing the server in `server_remove`.
Thus its IP won't be released until the end of the test that uses the
cluster and its logs and workdirs won't be cleared.

Note: a side effect of this change is that we no longer `uninstall`
servers in any other situation than suite cleanup (i.e. at the end of
the entire test run). There is a `ScyllaCluster.uninstall` function
which was probably intended to be used when we released a cluster (it
calls `ScyllaServer.uninstall`) but that function is not used anywhere,
not even when we get rid of a dirty cluster.