expr_test: Add basic unit tests for expr::evaluate()
expr::evaluate is a function used to evaluate
the value of an expression.

It's indirectly tested by using it in query
code, but there were no real unit tests for it.
Additionally some new features were added
recently (evaluating binary_operators),
for which testing by running queries isn't
sufficient.

Add unit tests for expr::evaluate.
They are not as exhaustive as they could be,
but they are enough for now.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>